requirements.txt
praw==7.8.1
rich==13.9.4
sqlalchemy==2.0.36
requests==2.32.3
python-dotenv==1.0.1
click==8.1.8 